# Backend/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangPlug API",
    "description": "German Language Learning Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Register",
        "description": "Register a new user",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "description": "Login user and create session",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "description": "Logout current user",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Current User Info",
        "description": "Get current user information",
        "operationId": "get_current_user_info_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/videos/user": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get User Videos",
        "description": "Get videos for the current user (alias for get_available_videos)",
        "operationId": "get_user_videos_api_videos_user_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get User Videos Api Videos User Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "description": "Get current user's profile",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/languages": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Supported Languages",
        "description": "Get list of supported languages",
        "operationId": "get_supported_languages_api_profile_languages_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response Get Supported Languages Api Profile Languages Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update Language Preferences",
        "description": "Update user's language preferences",
        "operationId": "update_language_preferences_api_profile_languages_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguagePreferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Language Preferences Api Profile Languages Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/settings": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get User Settings",
        "description": "Get user settings",
        "operationId": "get_user_settings_api_profile_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update User Settings",
        "description": "Update user settings",
        "operationId": "update_user_settings_api_profile_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/progress/user": {
      "get": {
        "tags": [
          "progress"
        ],
        "summary": "Get User Progress",
        "description": "Get progress data for the current user",
        "operationId": "get_user_progress_api_progress_user_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProgress"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token",
            "description": "JWT authentication token"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse",
            "description": "User information"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At",
            "description": "Token expiration timestamp (ISO format)"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user",
          "expires_at"
        ],
        "title": "AuthResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LanguagePreferences": {
        "properties": {
          "native_language": {
            "type": "string",
            "title": "Native Language"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          }
        },
        "type": "object",
        "required": [
          "native_language",
          "target_language"
        ],
        "title": "LanguagePreferences",
        "description": "Language preference update model"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username for authentication"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "Password for authentication"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Username",
            "description": "Username must be 3-50 characters, alphanumeric with underscores and hyphens only"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Password must be at least 8 characters long"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "UserProfile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_login": {
            "type": "string",
            "title": "Last Login"
          },
          "native_language": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Native Language"
          },
          "target_language": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Target Language"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "is_admin",
          "created_at",
          "native_language",
          "target_language"
        ],
        "title": "UserProfile",
        "description": "User profile response model"
      },
      "UserProgress": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "total_videos_watched": {
            "type": "integer",
            "title": "Total Videos Watched",
            "default": 0
          },
          "total_watch_time": {
            "type": "number",
            "title": "Total Watch Time",
            "default": 0.0
          },
          "vocabulary_learned": {
            "type": "integer",
            "title": "Vocabulary Learned",
            "default": 0
          },
          "current_streak": {
            "type": "integer",
            "title": "Current Streak",
            "default": 0
          },
          "longest_streak": {
            "type": "integer",
            "title": "Longest Streak",
            "default": 0
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": "Beginner"
          },
          "experience_points": {
            "type": "integer",
            "title": "Experience Points",
            "default": 0
          },
          "daily_goals_completed": {
            "type": "integer",
            "title": "Daily Goals Completed",
            "default": 0
          },
          "weekly_goals_completed": {
            "type": "integer",
            "title": "Weekly Goals Completed",
            "default": 0
          },
          "last_activity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity"
          },
          "achievements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          },
          "learning_stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Learning Stats",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserProgress",
        "description": "User progress model"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "description": "Unique user identifier"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "description": "Username"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "description": "Whether user has admin privileges"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether user account is active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Account creation timestamp (ISO format)"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login",
            "description": "Last login timestamp (ISO format)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "is_admin",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserSettings": {
        "properties": {
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme",
            "default": "light"
          },
          "notifications_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifications Enabled",
            "default": true
          },
          "auto_play": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Play",
            "default": true
          },
          "subtitle_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Size",
            "default": "medium"
          },
          "playback_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playback Speed",
            "default": 1.0
          },
          "vocabulary_difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocabulary Difficulty",
            "default": "intermediate"
          },
          "daily_goal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Goal",
            "default": 10
          },
          "language_preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguagePreferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UserSettings",
        "description": "User settings model"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}